+++ lib/src/carousel_controller/flutter_carousel_controller.dart
@@ -7,22 +7,31 @@ import '../enums/carousel_page_changed_reason.dart';
 import '../utils/flutter_carousel_utils.dart';
 
 abstract class FlutterCarouselController {
+  /// Factory constructor that returns an implementation of the controller.
   factory FlutterCarouselController() => FlutterCarouselControllerImpl();
 
+  /// Indicates if the controller is ready.
   bool get ready;
 
+  /// Future that completes when the controller is ready.
   Future<void> get onReady;
 
+  /// Navigates to the next page in the carousel.
   void nextPage({Duration? duration, Curve? curve});
 
+  /// Navigates to the previous page in the carousel.
   void previousPage({Duration? duration, Curve? curve});
 
+  /// Jumps to a specific page in the carousel without animation.
   void jumpToPage(int page);
 
+  /// Animates the carousel to a specific page.
   void animateToPage(int page, {Duration? duration, Curve? curve});
 
+  /// Starts the auto-play functionality of the carousel.
   void startAutoPlay();
 
+  /// Stops the auto-play functionality of the carousel.
   void stopAutoPlay();
 }
 
@@ -34,19 +43,24 @@ class FlutterCarouselControllerImpl implements FlutterCarouselController {
   /// The animation lasts for the given duration and follows the given curve.
   /// The returned [Future] resolves when the animation completes.
   @override
-  void animateToPage(int page,
-      {Duration? duration = const Duration(milliseconds: 300),
-      Curve? curve = Curves.linear}) async {
+  void animateToPage(
+    int page, {
+    Duration? duration = const Duration(milliseconds: 300),
+    Curve? curve = Curves.linear,
+  }) async {
     final isNeedResetTimer = _state!.options.pauseAutoPlayOnManualNavigate;
 
     if (isNeedResetTimer) {
       _state!.onResetTimer();
     }
+
+    // Calculate the real index to animate to.
     final index = getRealIndex(_state!.pageController?.page!.toInt() ?? 0,
         _state!.realPage - _state!.initialPage, _state!.itemCount);
 
     _setModeController();
 
+    // Animate to the target page.
     await _state!.pageController!.animateToPage(
         (_state!.pageController?.page!.toInt() ?? 0) + page - index,
         duration: duration!,
@@ -67,6 +81,7 @@ class FlutterCarouselControllerImpl implements FlutterCarouselController {
 
     _setModeController();
 
+    // Jump directly to the target page.
     return _state!.pageController?.jumpToPage(
         (_state!.pageController?.page!.toInt() ?? 0) + page - index);
   }
@@ -75,9 +90,10 @@ class FlutterCarouselControllerImpl implements FlutterCarouselController {
   /// The animation lasts for the given duration and follows the given curve.
   /// The returned [Future] resolves when the animation completes.
   @override
-  void nextPage(
-      {Duration? duration = const Duration(milliseconds: 300),
-      Curve? curve = Curves.linear}) async {
+  void nextPage({
+    Duration? duration = const Duration(milliseconds: 300),
+    Curve? curve = Curves.linear,
+  }) async {
     final isNeedResetTimer = _state!.options.pauseAutoPlayOnManualNavigate;
 
     if (isNeedResetTimer) {
@@ -86,6 +102,7 @@ class FlutterCarouselControllerImpl implements FlutterCarouselController {
 
     _setModeController();
 
+    // Navigate to the next page.
     await _state!.pageController?.nextPage(duration: duration!, curve: curve!);
 
     if (isNeedResetTimer) {
@@ -93,6 +110,7 @@ class FlutterCarouselControllerImpl implements FlutterCarouselController {
     }
   }
 
+  /// Future that completes when the controller is ready.
   @override
   Future<void> get onReady => _readyCompleter.future;
 
@@ -100,9 +118,10 @@ class FlutterCarouselControllerImpl implements FlutterCarouselController {
   /// The animation lasts for the given duration and follows the given curve.
   /// The returned [Future] resolves when the animation completes.
   @override
-  void previousPage(
-      {Duration? duration = const Duration(milliseconds: 300),
-      Curve? curve = Curves.linear}) async {
+  void previousPage({
+    Duration? duration = const Duration(milliseconds: 300),
+    Curve? curve = Curves.linear,
+  }) async {
     final isNeedResetTimer = _state!.options.pauseAutoPlayOnManualNavigate;
 
     if (isNeedResetTimer) {
@@ -111,6 +130,7 @@ class FlutterCarouselControllerImpl implements FlutterCarouselController {
 
     _setModeController();
 
+    // Navigate to the previous page.
     await _state!.pageController!
         .previousPage(duration: duration!, curve: curve!);
 
@@ -119,6 +139,7 @@ class FlutterCarouselControllerImpl implements FlutterCarouselController {
     }
   }
 
+  /// Checks if the controller is ready to interact with the carousel.
   @override
   bool get ready => _state != null;
 
@@ -138,6 +159,7 @@ class FlutterCarouselControllerImpl implements FlutterCarouselController {
     _state!.onResetTimer();
   }
 
+  /// Sets the current state of the carousel.
   set state(FlutterCarouselState? state) {
     _state = state;
     if (!_readyCompleter.isCompleted) {
@@ -145,6 +167,7 @@ class FlutterCarouselControllerImpl implements FlutterCarouselController {
     }
   }
 
+  /// Sets the carousel change mode to controller-based.
   void _setModeController() =>
       _state!.changeMode(CarouselPageChangedReason.controller);
 }