+++ lib/src/enums/center_page_enlarge_strategy.dart
@@ -1,1 +1,2 @@
+/// Enum for strategies to enlarge the center page in the carousel.
 enum CenterPageEnlargeStrategy { scale, height }