+++ lib/src/utils/flutter_carousel_utils.dart
@@ -1,5 +1,3 @@
-
-
 /// Calculates the actual index within a collection of items, considering the current
 /// position in an infinitely scrolling carousel.
 ///